+++ src/semidiscretization/semidiscretization_hyperbolic_parabolic.jl
@@ -39,46 +39,6 @@ struct SemidiscretizationHyperbolicParabolic{Mesh, Equations, EquationsParabolic
     cache_parabolic::CacheParabolic
 
     performance_counter::PerformanceCounterList{2}
-
-    function SemidiscretizationHyperbolicParabolic{Mesh, Equations, EquationsParabolic,
-                                                   InitialCondition, BoundaryConditions,
-                                                   BoundaryConditionsParabolic,
-                                                   SourceTerms, Solver,
-                                                   SolverParabolic, Cache,
-                                                   CacheParabolic}(mesh::Mesh,
-                                                                   equations::Equations,
-                                                                   equations_parabolic::EquationsParabolic,
-                                                                   initial_condition::InitialCondition,
-                                                                   boundary_conditions::BoundaryConditions,
-                                                                   boundary_conditions_parabolic::BoundaryConditionsParabolic,
-                                                                   source_terms::SourceTerms,
-                                                                   solver::Solver,
-                                                                   solver_parabolic::SolverParabolic,
-                                                                   cache::Cache,
-                                                                   cache_parabolic::CacheParabolic) where {
-                                                                                                           Mesh,
-                                                                                                           Equations,
-                                                                                                           EquationsParabolic,
-                                                                                                           InitialCondition,
-                                                                                                           BoundaryConditions,
-                                                                                                           BoundaryConditionsParabolic,
-                                                                                                           SourceTerms,
-                                                                                                           Solver,
-                                                                                                           SolverParabolic,
-                                                                                                           Cache,
-                                                                                                           CacheParabolic
-                                                                                                           }
-        @assert ndims(mesh) == ndims(equations)
-
-        # Todo: assert nvariables(equations)==nvariables(equations_parabolic)
-
-        performance_counter = PerformanceCounterList{2}(false)
-
-        new(mesh, equations, equations_parabolic, initial_condition,
-            boundary_conditions, boundary_conditions_parabolic,
-            source_terms, solver, solver_parabolic, cache, cache_parabolic,
-            performance_counter)
-    end
 end
 
 """
@@ -100,56 +60,53 @@ function SemidiscretizationHyperbolicParabolic(mesh, equations::Tuple,
                                                # `RealT` is used as real type for node locations etc.
                                                # while `uEltype` is used as element type of solutions etc.
                                                RealT = real(solver), uEltype = RealT)
-    equations_hyperbolic, equations_parabolic = equations
-    boundary_conditions_hyperbolic, boundary_conditions_parabolic = boundary_conditions
+    equations, equations_parabolic = equations
+    boundary_conditions, boundary_conditions_parabolic = boundary_conditions
 
-    return SemidiscretizationHyperbolicParabolic(mesh, equations_hyperbolic,
-                                                 equations_parabolic,
-                                                 initial_condition, solver;
-                                                 solver_parabolic, source_terms,
-                                                 boundary_conditions = boundary_conditions_hyperbolic,
-                                                 boundary_conditions_parabolic = boundary_conditions_parabolic,
-                                                 RealT, uEltype)
-end
+    @assert ndims(mesh) == ndims(equations)
+    @assert ndims(mesh) == ndims(equations_parabolic)
 
-function SemidiscretizationHyperbolicParabolic(mesh, equations, equations_parabolic,
-                                               initial_condition, solver;
-                                               solver_parabolic = default_parabolic_solver(),
-                                               source_terms = nothing,
-                                               boundary_conditions = boundary_condition_periodic,
-                                               boundary_conditions_parabolic = boundary_condition_periodic,
-                                               # `RealT` is used as real type for node locations etc.
-                                               # while `uEltype` is used as element type of solutions etc.
-                                               RealT = real(solver), uEltype = RealT)
+    if !(nvariables(equations) == nvariables(equations_parabolic))
+        throw(ArgumentError("Current implementation of viscous terms requires the same number of conservative and gradient variables."))
+    end
+
     cache = create_cache(mesh, equations, solver, RealT, uEltype)
-    _boundary_conditions = digest_boundary_conditions(boundary_conditions, mesh, solver,
-                                                      cache)
-    _boundary_conditions_parabolic = digest_boundary_conditions(boundary_conditions_parabolic,
-                                                                mesh, solver, cache)
-
+    _boundary_conditions = digest_boundary_conditions(boundary_conditions,
+                                                      mesh, solver, cache)
     check_periodicity_mesh_boundary_conditions(mesh, _boundary_conditions)
 
     cache_parabolic = create_cache_parabolic(mesh, equations, equations_parabolic,
-                                             solver, solver_parabolic, RealT,
-                                             uEltype)
+                                             solver, solver_parabolic,
+                                             RealT, uEltype)
 
-    SemidiscretizationHyperbolicParabolic{typeof(mesh), typeof(equations),
-                                          typeof(equations_parabolic),
-                                          typeof(initial_condition),
-                                          typeof(_boundary_conditions),
-                                          typeof(_boundary_conditions_parabolic),
-                                          typeof(source_terms), typeof(solver),
-                                          typeof(solver_parabolic), typeof(cache),
-                                          typeof(cache_parabolic)}(mesh, equations,
-                                                                   equations_parabolic,
-                                                                   initial_condition,
-                                                                   _boundary_conditions,
-                                                                   _boundary_conditions_parabolic,
-                                                                   source_terms,
-                                                                   solver,
-                                                                   solver_parabolic,
-                                                                   cache,
-                                                                   cache_parabolic)
+    _boundary_conditions_parabolic = digest_boundary_conditions(boundary_conditions_parabolic,
+                                                                mesh, solver, cache)
+    check_periodicity_mesh_boundary_conditions(mesh, _boundary_conditions_parabolic)
+
+    performance_counter = PerformanceCounterList{2}(false)
+
+    return SemidiscretizationHyperbolicParabolic{typeof(mesh),
+                                                 typeof(equations),
+                                                 typeof(equations_parabolic),
+                                                 typeof(initial_condition),
+                                                 typeof(_boundary_conditions),
+                                                 typeof(_boundary_conditions_parabolic),
+                                                 typeof(source_terms),
+                                                 typeof(solver),
+                                                 typeof(solver_parabolic),
+                                                 typeof(cache),
+                                                 typeof(cache_parabolic)}(mesh,
+                                                                          equations,
+                                                                          equations_parabolic,
+                                                                          initial_condition,
+                                                                          _boundary_conditions,
+                                                                          _boundary_conditions_parabolic,
+                                                                          source_terms,
+                                                                          solver,
+                                                                          solver_parabolic,
+                                                                          cache,
+                                                                          cache_parabolic,
+                                                                          performance_counter)
 end
 
 # Create a new semidiscretization but change some parameters compared to the input.
@@ -171,10 +128,12 @@ function remake(semi::SemidiscretizationHyperbolicParabolic;
     # TODO: Which parts do we want to `remake`? At least the solver needs some
     #       special care if shock-capturing volume integrals are used (because of
     #       the indicators and their own caches...).
-    SemidiscretizationHyperbolicParabolic(mesh, equations, equations_parabolic,
-                                          initial_condition, solver; solver_parabolic,
-                                          source_terms, boundary_conditions,
-                                          boundary_conditions_parabolic, uEltype)
+    return SemidiscretizationHyperbolicParabolic(mesh, (equations, equations_parabolic),
+                                                 initial_condition, solver;
+                                                 solver_parabolic, source_terms,
+                                                 boundary_conditions = (boundary_conditions,
+                                                                        boundary_conditions_parabolic),
+                                                 uEltype)
 end
 
 function Base.show(io::IO, semi::SemidiscretizationHyperbolicParabolic)
@@ -264,15 +223,28 @@ function get_node_variables!(node_variables, u_ode,
 end
 
 """
-    semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan)
+    semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan;
+                   jac_prototype_parabolic::Union{AbstractMatrix, Nothing} = nothing,
+                   colorvec_parabolic::Union{AbstractVector, Nothing} = nothing)
 
 Wrap the semidiscretization `semi` as a split ODE problem in the time interval `tspan`
 that can be passed to `solve` from the [SciML ecosystem](https://diffeq.sciml.ai/latest/).
 The parabolic right-hand side is the first function of the split ODE problem and
 will be used by default by the implicit part of IMEX methods from the
 SciML ecosystem.
+
+Optional keyword arguments:
+- `jac_prototype_parabolic`: Expected to come from [SparseConnectivityTracer.jl](https://github.com/adrhill/SparseConnectivityTracer.jl).
+  Specifies the sparsity structure of the parabolic function's Jacobian to enable e.g. efficient implicit time stepping.
+  The [`SplitODEProblem`](https://docs.sciml.ai/DiffEqDocs/stable/types/split_ode_types/#SciMLBase.SplitODEProblem) only expects the Jacobian
+  to be defined on the first function it takes in, which is treated implicitly. This corresponds to the parabolic right-hand side in Trixi.jl.
+  The hyperbolic right-hand side is expected to be treated explicitly, and therefore its Jacobian is irrelevant.
+- `colorvec_parabolic`: Expected to come from [SparseMatrixColorings.jl](https://github.com/gdalle/SparseMatrixColorings.jl).
+  Allows for even faster Jacobian computation. Not necessarily required when `jac_prototype_parabolic` is given.
 """
 function semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan;
+                        jac_prototype_parabolic::Union{AbstractMatrix, Nothing} = nothing,
+                        colorvec_parabolic::Union{AbstractVector, Nothing} = nothing,
                         reset_threads = true)
     # Optionally reset Polyester.jl threads. See
     # https://github.com/trixi-framework/Trixi.jl/issues/1583
@@ -286,15 +258,33 @@ function semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan;
     #       mpi_isparallel() && MPI.Barrier(mpi_comm())
     #       See https://github.com/trixi-framework/Trixi.jl/issues/328
     iip = true # is-inplace, i.e., we modify a vector when calling rhs_parabolic!, rhs!
-    # Note that the IMEX time integration methods of OrdinaryDiffEq.jl treat the
-    # first function implicitly and the second one explicitly. Thus, we pass the
-    # stiffer parabolic function first.
-    return SplitODEProblem{iip}(rhs_parabolic!, rhs!, u0_ode, tspan, semi)
+
+    # Check if Jacobian prototype is provided for sparse Jacobian
+    if jac_prototype_parabolic !== nothing
+        # Convert `jac_prototype_parabolic` to real type, as seen here:
+        # https://docs.sciml.ai/DiffEqDocs/stable/tutorials/advanced_ode_example/#Declaring-a-Sparse-Jacobian-with-Automatic-Sparsity-Detection
+        parabolic_ode = SciMLBase.ODEFunction(rhs_parabolic!,
+                                              jac_prototype = convert.(eltype(u0_ode),
+                                                                       jac_prototype_parabolic),
+                                              colorvec = colorvec_parabolic) # coloring vector is optional
+
+        # Note that the IMEX time integration methods of OrdinaryDiffEq.jl treat the
+        # first function implicitly and the second one explicitly. Thus, we pass the
+        # (potentially) stiffer parabolic function first.
+        return SplitODEProblem{iip}(parabolic_ode, rhs!, u0_ode, tspan, semi)
+    else
+        # We could also construct an `ODEFunction` explicitly without the Jacobian here,
+        # but we stick to the lean direct in-place functions `rhs_parabolic!` and
+        # let OrdinaryDiffEq.jl handle the rest
+        return SplitODEProblem{iip}(rhs_parabolic!, rhs!, u0_ode, tspan, semi)
+    end
 end
 
 """
     semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan,
-                   restart_file::AbstractString)
+                   restart_file::AbstractString;
+                   jac_prototype_parabolic::Union{AbstractMatrix, Nothing} = nothing,
+                   colorvec_parabolic::Union{AbstractVector, Nothing} = nothing)
 
 Wrap the semidiscretization `semi` as a split ODE problem in the time interval `tspan`
 that can be passed to `solve` from the [SciML ecosystem](https://diffeq.sciml.ai/latest/).
@@ -303,9 +293,20 @@ will be used by default by the implicit part of IMEX methods from the
 SciML ecosystem.
 
 The initial condition etc. is taken from the `restart_file`.
+
+Optional keyword arguments:
+- `jac_prototype_parabolic`: Expected to come from [SparseConnectivityTracer.jl](https://github.com/adrhill/SparseConnectivityTracer.jl).
+  Specifies the sparsity structure of the parabolic function's Jacobian to enable e.g. efficient implicit time stepping.
+  The [`SplitODEProblem`](https://docs.sciml.ai/DiffEqDocs/stable/types/split_ode_types/#SciMLBase.SplitODEProblem) only expects the Jacobian
+  to be defined on the first function it takes in, which is treated implicitly. This corresponds to the parabolic right-hand side in Trixi.jl.
+  The hyperbolic right-hand side is expected to be treated explicitly, and therefore its Jacobian is irrelevant.
+- `colorvec_parabolic`: Expected to come from [SparseMatrixColorings.jl](https://github.com/gdalle/SparseMatrixColorings.jl).
+  Allows for even faster Jacobian computation. Not necessarily required when `jac_prototype_parabolic` is given.
 """
 function semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan,
                         restart_file::AbstractString;
+                        jac_prototype_parabolic::Union{AbstractMatrix, Nothing} = nothing,
+                        colorvec_parabolic::Union{AbstractVector, Nothing} = nothing,
                         reset_threads = true)
     # Optionally reset Polyester.jl threads. See
     # https://github.com/trixi-framework/Trixi.jl/issues/1583
@@ -319,10 +320,26 @@ function semidiscretize(semi::SemidiscretizationHyperbolicParabolic, tspan,
     #       mpi_isparallel() && MPI.Barrier(mpi_comm())
     #       See https://github.com/trixi-framework/Trixi.jl/issues/328
     iip = true # is-inplace, i.e., we modify a vector when calling rhs_parabolic!, rhs!
-    # Note that the IMEX time integration methods of OrdinaryDiffEq.jl treat the
-    # first function implicitly and the second one explicitly. Thus, we pass the
-    # stiffer parabolic function first.
-    return SplitODEProblem{iip}(rhs_parabolic!, rhs!, u0_ode, tspan, semi)
+
+    # Check if Jacobian prototype is provided for sparse Jacobian
+    if jac_prototype_parabolic !== nothing
+        # Convert `jac_prototype_parabolic` to real type, as seen here:
+        # https://docs.sciml.ai/DiffEqDocs/stable/tutorials/advanced_ode_example/#Declaring-a-Sparse-Jacobian-with-Automatic-Sparsity-Detection
+        parabolic_ode = SciMLBase.ODEFunction(rhs_parabolic!,
+                                              jac_prototype = convert.(eltype(u0_ode),
+                                                                       jac_prototype_parabolic),
+                                              colorvec = colorvec_parabolic) # coloring vector is optional
+
+        # Note that the IMEX time integration methods of OrdinaryDiffEq.jl treat the
+        # first function implicitly and the second one explicitly. Thus, we pass the
+        # (potentially) stiffer parabolic function first.
+        return SplitODEProblem{iip}(parabolic_ode, rhs!, u0_ode, tspan, semi)
+    else
+        # We could also construct an `ODEFunction` explicitly without the Jacobian here,
+        # but we stick to the lean direct in-place function `rhs_parabolic!` and
+        # let OrdinaryDiffEq.jl handle the rest
+        return SplitODEProblem{iip}(rhs_parabolic!, rhs!, u0_ode, tspan, semi)
+    end
 end
 
 function rhs!(du_ode, u_ode, semi::SemidiscretizationHyperbolicParabolic, t)
@@ -361,6 +378,64 @@ function rhs_parabolic!(du_ode, u_ode, semi::SemidiscretizationHyperbolicParabol
     return nothing
 end
 
+"""
+    linear_structure(semi::SemidiscretizationHyperbolicParabolic;
+                     t0 = zero(real(semi)))
+
+Wraps the right-hand side operator of the hyperbolic-parabolic semidiscretization `semi`
+at time `t0` as an affine-linear operator given by a linear operator `A`
+and a vector `b`:
+```math
+\\partial_t u(t) = A u(t) - b.
+```
+Works only for linear equations, i.e., equations with `have_constant_speed(equations) == True()`.
+
+This has the benefit of greatly reduced memory consumption compared to constructing
+the full system matrix explicitly, as done for instance in
+[`jacobian_fd`](@ref) and [`jacobian_ad_forward`](@ref).
+
+The returned linear operator `A` is a matrix-free representation which can be
+supplied to iterative solvers from, e.g., [Krylov.jl](https://github.com/JuliaSmoothOptimizers/Krylov.jl).
+"""
+function linear_structure(semi::SemidiscretizationHyperbolicParabolic;
+                          t0 = zero(real(semi)))
+    if have_constant_speed(semi.equations) == False()
+        throw(ArgumentError("`linear_structure` expects linear equations."))
+    end
+
+    # allocate memory
+    u_ode = allocate_coefficients(mesh_equations_solver_cache(semi)...)
+    du_ode = similar(u_ode)
+
+    # get the right hand side from boundary conditions and optional source terms
+    u_ode .= zero(eltype(u_ode))
+    rhs!(du_ode, u_ode, semi, t0)
+    b = -du_ode
+
+    # Repeat for parabolic part
+    rhs_parabolic!(du_ode, u_ode, semi, t0)
+    @. b -= du_ode
+
+    # Create a copy of `b` used internally to extract the linear part of `semi`.
+    # This is necessary to get everything correct when the user updates the
+    # returned vector `b`.
+    b_tmp = copy(b)
+
+    # additional storage for parabolic part
+    dest_para = similar(du_ode)
+
+    # wrap the linear operator
+    A = LinearMap(length(u_ode), ismutating = true) do dest, src
+        rhs!(dest, src, semi, t0)
+        rhs_parabolic!(dest_para, src, semi, t0)
+
+        @. dest += dest_para + b_tmp
+        return dest
+    end
+
+    return A, b
+end
+
 function _jacobian_ad_forward(semi::SemidiscretizationHyperbolicParabolic, t0, u0_ode,
                               du_ode, config)
     new_semi = remake(semi, uEltype = eltype(config))
@@ -378,8 +453,8 @@ end
 
 """
     jacobian_ad_forward_parabolic(semi::SemidiscretizationHyperbolicParabolic;
-                                  t0=zero(real(semi)),
-                                  u0_ode=compute_coefficients(t0, semi))
+                                  t0 = zero(real(semi)),
+                                  u0_ode = compute_coefficients(t0, semi))
 
 Uses the *parabolic part* of the right-hand side operator of the [`SemidiscretizationHyperbolicParabolic`](@ref) `semi`
 and forward mode automatic differentiation to compute the Jacobian `J` of the