+++ src/solvers/dgsem_tree/containers.jl
@@ -8,6 +8,70 @@
 # Dimension independent code related to containers of the DG solver
 # with the mesh type TreeMesh
 
+abstract type AbstractTreeElementContainer <: AbstractElementContainer end
+
+# Return number of elements
+@inline nelements(elements::AbstractTreeElementContainer) = length(elements.cell_ids)
+# Return number of element nodes
+@inline nnodes(elements::AbstractTreeElementContainer) = size(elements.node_coordinates,
+                                                              2)
+@inline nvariables(elements::AbstractTreeElementContainer) = size(elements.surface_flux_values,
+                                                                  1)
+# TODO: Taal performance, 1:nelements(elements) vs. Base.OneTo(nelements(elements))
+"""
+    eachelement(elements::AbstractTreeElementContainer)
+
+Return an iterator over the indices that specify the location in relevant data structures
+for the elements in `elements`. 
+In particular, not the elements themselves are returned.
+"""
+@inline eachelement(elements::AbstractTreeElementContainer) = Base.OneTo(nelements(elements))
+
+@inline Base.real(elements::AbstractTreeElementContainer) = eltype(elements.node_coordinates)
+@inline Base.eltype(elements::AbstractTreeElementContainer) = eltype(elements.surface_flux_values)
+
+abstract type AbstractTreeInterfaceContainer <: AbstractInterfaceContainer end
+
+# Return number of interfaces
+@inline ninterfaces(interfaces::AbstractTreeInterfaceContainer) = length(interfaces.orientations)
+# Return number of interface nodes for 2D and 3D. For 1D hard-coded to 1 interface node.
+@inline nnodes(interfaces::AbstractTreeInterfaceContainer) = size(interfaces.u, 3)
+# Return number of equation variables
+@inline nvariables(interfaces::AbstractTreeInterfaceContainer) = size(interfaces.u, 2)
+
+abstract type AbstractTreeMPIInterfaceContainer <: AbstractMPIInterfaceContainer end
+
+# Return number of interfaces
+@inline function nmpiinterfaces(mpi_interfaces::AbstractTreeMPIInterfaceContainer)
+    return length(mpi_interfaces.orientations)
+end
+# Return number of interface nodes for 2D and 3D. For 1D hard-coded to 1 interface node.
+@inline nnodes(interfaces::AbstractTreeMPIInterfaceContainer) = size(interfaces.u, 3)
+# Return number of equation variables
+@inline nvariables(interfaces::AbstractTreeMPIInterfaceContainer) = size(interfaces.u,
+                                                                         2)
+
+abstract type AbstractTreeBoundaryContainer <: AbstractBoundaryContainer end
+
+# Return number of boundaries
+@inline nboundaries(boundaries::AbstractTreeBoundaryContainer) = length(boundaries.orientations)
+# Return number of boundary nodes for 2D and 3D. For 1D hard-coded to 1 boundary node.
+@inline nnodes(boundaries::AbstractTreeBoundaryContainer) = size(boundaries.u, 3)
+# Return number of equation variables
+@inline nvariables(boundaries::AbstractTreeBoundaryContainer) = size(boundaries.u, 2)
+
+abstract type AbstractTreeL2MortarContainer <: AbstractMortarContainer end
+
+# Return number of L2 mortars
+@inline nmortars(l2mortars::AbstractTreeL2MortarContainer) = length(l2mortars.orientations)
+
+abstract type AbstractTreeL2MPIMortarContainer <: AbstractMPIMortarContainer end
+
+# Return number of L2 mortars
+@inline function nmpimortars(mpi_l2mortars::AbstractTreeL2MPIMortarContainer)
+    length(mpi_l2mortars.orientations)
+end
+
 function reinitialize_containers!(mesh::TreeMesh, equations, dg::DGSEM, cache)
     # Get new list of leaf cells
     leaf_cell_ids = local_leaf_cells(mesh.tree)
@@ -118,6 +182,12 @@ function Base.resize!(container::ContainerSubcellLimiterIDP, capacity)
 
     return nothing
 end
+
+abstract type AbstractContainerAntidiffusiveFlux <: AbstractContainer end
+nvariables(fluxes::AbstractContainerAntidiffusiveFlux) = size(fluxes.antidiffusive_flux1_L,
+                                                              1)
+nnodes(fluxes::AbstractContainerAntidiffusiveFlux) = size(fluxes.antidiffusive_flux1_L,
+                                                          3)
 
 # Dimension-specific implementations
 include("containers_1d.jl")